Moved capacity prediction modules to own packages to simplify project
