+++ stf/model/capacity/predict.py
@@ -10,9 +10,9 @@ from ktpbase.log import logging
 from stf.feature_engineering.capacity_prognoses_features import (
     apply_capacity_features
 )
-from stf.model.capacity_prognosis_model import CapacityPrognosisModel
-from stf.model.capacity_prognosis_utils import (
-    prepare_prediction_data, prepare_training_data, visualize_predictions
+from stf.model.capacity.model import CapacityPredictionModel
+from stf.model.capacity.utils import (
+    prepare_prediction_data, visualize_predictions
 )
 from stf.model.serializer.xgboost.xgboost import XGBModelSerializer
 
@@ -55,12 +55,11 @@ def predict_capacity_prognosis(
 
     # create model
     logger.info("Create capacity prognosis model")
-    model = CapacityPrognosisModel()
+    model = CapacityPredictionModel()
 
     # load model
     serializer = XGBModelSerializer()
-    model_location = serializer.trained_models_folder
-    directory = model_location / str(pj["id"]) / "capacity"
+    directory = serializer.trained_models_folder / str(pj["id"]) / "capacity"
     logger.info("Loading model", model_directory=directory)
     model.load(directory=directory)
 
@@ -80,62 +79,3 @@ def predict_capacity_prognosis(
         auto_open=False,
     )
     logger.info("Sucessfully made capacity prediction")
-
-
-def train_capacity_prognosis(pj, datetime_start, datetime_end, y_hor=[0, 6, 13]):
-    """ Train a capacity prognoses model for a specific prediction job.
-
-    Args:
-        pj: (dict) prediction job
-    """
-    db = DataBase()
-
-    # initialize logging
-    logger = logging.get_logger(__name__)
-
-    # get input data
-    logger.info("Get input data")
-    load_data = db.get_load_pid(
-        pid=pj["id"], datetime_start=datetime_start, datetime_end=datetime_end
-    )
-    load_profiles_data = db.get_tdcv_load_profiles(datetime_start, datetime_end)
-    load_profile_names = list(load_profiles_data.columns)
-
-    input_data = pd.concat([load_data, load_profiles_data], axis=1)
-
-    # apply features
-    logger.info("Apply features")
-    feature_data, classes = apply_capacity_features(
-        input_data, y_col="load_max", y_hor=y_hor, outlier_removal=False,
-        load_profile_names=load_profile_names
-    )
-
-    # prepare data
-    logger.info("Prepare prediction data")
-    train_x, train_y, train_h, val_x, val_y, val_h = prepare_training_data(
-        feature_data, y_col="load_max"
-    )
-
-    # create model
-    logger.info("Create capacity prognosis model")
-    model = CapacityPrognosisModel(classes=classes)
-
-    # train model
-    logger.info("Train model")
-    model.train(train_x, train_y, val_x, val_y)
-
-    # evaluate model performance
-    logger.info("Evaluate model")
-    scores = {
-        "train": model.evaluate(train_x, train_y),
-        "validation": model.evaluate(val_x, val_y),
-    }  # Gives code smell but is necessary for unit tests
-    # TODO: make train report?
-    logger.info("Training metrics", scores=scores)
-
-    # save model
-    logger.info("Saving model")
-    serializer = XGBModelSerializer()
-    model_location = str(serializer.trained_models_folder)
-    directory = model_location + "/" + str(pj["id"]) + "/capacity/"
-    model.save(directory=directory)
