+++ stf/model/capacity/utils.py
